Move IntegrationTestRepoConfig class to another module
